server/ml/odai/requirements.txt
pydantic
google-cloud-storage
python-dotenv
